move to qis subpackage
