+++ cirq/ops/linear_combinations.py
@@ -18,7 +18,7 @@ import numbers
 
 import numpy as np
 
-from cirq import linalg, protocols, value
+from cirq import protocols, qis, value
 from cirq._doc import document
 from cirq.linalg import operator_spaces
 from cirq.ops import identity, raw_types, pauli_gates, pauli_string
@@ -392,10 +392,10 @@ class PauliSum:
                              "with shape `(2 ** n,)` or `(2, ..., 2)`.")
 
         if check_preconditions:
-            linalg.validate_normalized_state(state=state,
-                                             qid_shape=(2,) * num_qubits,
-                                             dtype=state.dtype,
-                                             atol=atol)
+            qis.validate_normalized_state(state=state,
+                                          qid_shape=(2,) * num_qubits,
+                                          dtype=state.dtype,
+                                          atol=atol)
         return sum(
             p._expectation_from_wavefunction_no_validation(state, qubit_map)
             for p in self)
@@ -444,11 +444,11 @@ class PauliSum:
 
         if check_preconditions:
             # Do not enforce reshaping if the state all axes are dimension 2.
-            _ = linalg.to_valid_density_matrix(density_matrix_rep=state.reshape(
+            _ = qis.to_valid_density_matrix(density_matrix_rep=state.reshape(
                 dim, dim),
-                                               num_qubits=num_qubits,
-                                               dtype=state.dtype,
-                                               atol=atol)
+                                            num_qubits=num_qubits,
+                                            dtype=state.dtype,
+                                            atol=atol)
         return sum(
             p._expectation_from_density_matrix_no_validation(state, qubit_map)
             for p in self)
