+++ cirq/protocols/unitary_protocol.py
@@ -23,7 +23,7 @@ from typing import (
 import numpy as np
 from typing_extensions import Protocol
 
-from cirq import linalg
+from cirq import qis
 from cirq._doc import document
 from cirq.protocols import qid_shape_protocol
 from cirq.protocols.apply_unitary_protocol import (
@@ -172,7 +172,7 @@ def _strat_unitary_from_apply_unitary(val: Any) -> Optional[np.ndarray]:
         return NotImplemented
 
     # Apply unitary effect to an identity matrix.
-    state = linalg.eye_tensor(val_qid_shape, dtype=np.complex128)
+    state = qis.eye_tensor(val_qid_shape, dtype=np.complex128)
     buffer = np.empty_like(state)
     result = method(ApplyUnitaryArgs(state, buffer, range(len(val_qid_shape))))
 
@@ -191,7 +191,7 @@ def _strat_unitary_from_decompose(val: Any) -> Optional[np.ndarray]:
         return NotImplemented
 
     # Apply sub-operations' unitary effects to an identity matrix.
-    state = linalg.eye_tensor(val_qid_shape, dtype=np.complex128)
+    state = qis.eye_tensor(val_qid_shape, dtype=np.complex128)
     buffer = np.empty_like(state)
     result = apply_unitaries(
         operations, qubits,
