+++ cirq/sim/density_matrix_utils_test.py
@@ -18,6 +18,21 @@ import numpy as np
 import pytest
 
 import cirq
+from cirq._compat_test import capture_logging
+
+
+def test_deprecated():
+    with capture_logging() as log:
+        _ = cirq.sim.to_valid_density_matrix(0, 1)
+    assert len(log) == 1
+    assert "cirq.to_valid_density_matrix" in log[0].getMessage()
+    assert "deprecated" in log[0].getMessage()
+
+    with capture_logging() as log:
+        _ = cirq.sim.von_neumann_entropy(np.eye(2) / 2)
+    assert len(log) == 1
+    assert "cirq.von_neumann_entropy" in log[0].getMessage()
+    assert "deprecated" in log[0].getMessage()
 
 
 def test_sample_density_matrix_big_endian():
